move a few things around
