+++ yarn-project/simulator/src/index.ts
@@ -3,7 +3,7 @@ export * from './avm/index.js';
 export * from './client/index.js';
 export * from './common/index.js';
 export * from './public/index.js';
-export * from './simulator/index.js';
+export * from './providers/index.js';
 export * from './mocks/index.js';
 export * from './stats/index.js';
 export * from './utils.js';
