+++ yarn-project/simulator/src/providers/acvm_native.ts
