+++ yarn-project/simulator/src/providers/acvm_wasm.ts
